+++ ProyectoP2/DTOs/VentaDTO.cs
@@ -5,9 +5,9 @@ namespace ProyectoP2.DTOs
     {
         public string Cliente { get; set; }
         public string NumeroVenta { get; set; }
-        public decimal Total { get; set; }
-        public decimal PagoCon { get; set; }
-        public decimal Cambio { get; set; }
+        public double Total { get; set; }
+        public double PagoCon { get; set; }
+        public double Cambio { get; set; }
         public string FechaRegistro { get; set; }
         public List<DetalleVentaDTO> DetalleVenta { get; set; } = new List<DetalleVentaDTO>();
     }
