+++ ProyectoP2/Models/Venta.cs
@@ -13,9 +13,9 @@ namespace ProyectoP2.Models
         public int IdVenta { get; set; }
         public string Cliente { get; set; }
         public string NumeroVenta { get; set; }
-        public decimal Total { get; set; }
-        public decimal PagoCon { get; set; }
-        public decimal Cambio { get; set; }
+        public double Total { get; set; }
+        public double PagoCon { get; set; }
+        public double Cambio { get; set; }
         public DateTime FechaRegistro { get; set; }
         public virtual ICollection<DetalleVenta> RefDetalleVenta { get; set; } = new List<DetalleVenta>();
     }
